refactor: renamed to follow JS standards
